[PATCH] Make _untar_bin use Unix-syntax archive names on VMS.
There are two tar utilities likely to be found on VMS systems. One
is a port of GNU tar that does not understand native-syntax
filenames.  The other is called vmstar and it understands either,
so the safest thing seems to be to always specify the archive
name in Unix format.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>