Correctly detect lack of /bin/tar
The higher-precedence not was applying to the expression as a whole,
rather than the first component.

This was causing breakage when using $PREFER_BIN on a system without a tar
command (such as Windows typically is), such as in the installer for
Alien::wxWidgets.

The high-precedence first meant that the check for the tar command was
passing when it should fail, and second that a later failing check (for
example when gzip also isn't found) had its error message inverted to the
empty string.