Refactor type detection into type_for() function
Should be useful for situations in which the archive file does not contain an
extension, but one has a separate string containing a file name and extension.
This comes up for file uploads to a web server, for example, where the server
implementation writes the file with a random temporary name and provides the
original file name in a separate field.